Add react-cosmos to simplify visual component testing (#452)
